Modify PKO send to use an atomic increment/decrement for controlling queue depth. This is to workaround performannce problems and crashes that seem to be caused by accesses to PKO.
git-svn-id: http://casvn/svn/octeon/bdk/trunk@98321 c83883e8-9f28-0410-baf1-f2025811e16a

+++ libbdk-hal/if/bdk-if.h
@@ -78,7 +78,7 @@ typedef struct __bdk_if_port
     int         pko_queue   : 16;   /* CN78XX has 1024, CN6XXX has 256 */
     int         aura        : 16;   /* CN78XX has 1024, CN6XXX doesn't use */
     bdk_if_flags_t flags    : 8;
-    int         pko_depth;          /* Tracks the PKO DQ depth on 78. This is a workaround for slow PKO performance */
+    int64_t     pko_depth;          /* Tracks the PKO DQ depth on 78, decremented by hardware. This is a workaround for slow PKO performance */
     void *      receiver;           /* This is a bdk_if_packet_receiver_t */
     void *      receiver_arg;
     void *      priv;
